pyproject.toml
[tool.poetry]
name = "penta-event-management"
version = "0.1.0"
description = "Event management"
authors = ["Nazmul <nazmulfarhan@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.112.0"
sqlalchemy = "^2.0.32"
sqlalchemy-utils = "^0.41.2"
python-dotenv = "^1.0.1"
asyncpg = "^0.29.0"
uvicorn = "^0.30.6"
pydantic = "^2.8.2"
tenacity = "^9.0.0"
sqlalchemy-json = "^0.7.0"
aiohttp = "^3.10.4"
dependency-injector = "4.41.0"
alembic = "^1.13.2"
containers = "^0.0.4"
pyjwt = "^2.9.0"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.9"
python-jose = "^3.3.0"
fastapi-keycloak-middleware = "^1.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[virtualenvs]
create = true
in-project = true
